feat(database): change the data type of the target and amount attributes

+++ database/migrations/2024_12_28_133451_create_programs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->string('name');
             $table->longText('description');
-            $table->decimal('target', 10, 2);
+            $table->bigInteger('target');
             $table->timestamps();
             $table->softDeletes();
         });
